requirements.txt
python-telegram-bot==21.6
openpyxl==3.1.2
gspread==6.0.2
google-auth==2.28.1
pytz==2024.1
python-dotenv==1.0.1
